+++ hutool-http/src/main/java/cn/hutool/http/HttpDownloader.java
@@ -11,8 +11,8 @@ import java.nio.charset.Charset;
 /**
  * 下载封装，下载统一使用{@code GET}请求，默认支持30x跳转
  *
- * @since 5.6.4
  * @author looly
+ * @since 5.6.4
  */
 public class HttpDownloader {
 
@@ -43,27 +43,43 @@ public class HttpDownloader {
 	/**
 	 * 下载远程文件
 	 *
-	 * @param url            请求的url
-	 * @param destFile       目标文件或目录，当为目录时，取URL中的文件名，取不到使用编码后的URL做为文件名
-	 * @param timeout        超时，单位毫秒，-1表示默认超时
-	 * @param streamProgress 进度条
+	 * @param url             请求的url
+	 * @param targetFileOrDir 目标文件或目录，当为目录时，取URL中的文件名，取不到使用编码后的URL做为文件名
+	 * @param timeout         超时，单位毫秒，-1表示默认超时
+	 * @param streamProgress  进度条
 	 * @return 文件大小
 	 */
-	public static long downloadFile(String url, File destFile, int timeout, StreamProgress streamProgress) {
-		return requestDownload(url, timeout).writeBody(destFile, streamProgress);
+	public static long downloadFile(String url, File targetFileOrDir, int timeout, StreamProgress streamProgress) {
+		return requestDownload(url, timeout).writeBody(targetFileOrDir, streamProgress);
+	}
+
+	/**
+	 * 下载文件-避免未完成的文件<br>
+	 * 来自：https://gitee.com/dromara/hutool/pulls/407<br>
+	 * 此方法原理是先在目标文件同级目录下创建临时文件，下载之，等下载完毕后重命名，避免因下载错误导致的文件不完整。
+	 *
+	 * @param url             请求的url
+	 * @param targetFileOrDir 目标文件或目录，当为目录时，取URL中的文件名，取不到使用编码后的URL做为文件名
+	 * @param tempFileSuffix  临时文件后缀，默认".temp"
+	 * @param timeout         超时，单位毫秒，-1表示默认超时
+	 * @param streamProgress  进度条
+	 * @since 5.7.12
+	 */
+	public long downloadFile(String url, File targetFileOrDir, String tempFileSuffix, int timeout, StreamProgress streamProgress) {
+		return requestDownload(url, timeout).writeBody(targetFileOrDir, tempFileSuffix, streamProgress);
 	}
 
 	/**
 	 * 下载远程文件，返回文件
 	 *
-	 * @param url            请求的url
-	 * @param destFile       目标文件或目录，当为目录时，取URL中的文件名，取不到使用编码后的URL做为文件名
-	 * @param timeout        超时，单位毫秒，-1表示默认超时
-	 * @param streamProgress 进度条
+	 * @param url             请求的url
+	 * @param targetFileOrDir 目标文件或目录，当为目录时，取URL中的文件名，取不到使用编码后的URL做为文件名
+	 * @param timeout         超时，单位毫秒，-1表示默认超时
+	 * @param streamProgress  进度条
 	 * @return 文件
 	 */
-	public static File downloadForFile(String url, File destFile, int timeout, StreamProgress streamProgress) {
-		return requestDownload(url, timeout).writeBodyForFile(destFile, streamProgress);
+	public static File downloadForFile(String url, File targetFileOrDir, int timeout, StreamProgress streamProgress) {
+		return requestDownload(url, timeout).writeBodyForFile(targetFileOrDir, streamProgress);
 	}
 
 	/**
@@ -84,8 +100,8 @@ public class HttpDownloader {
 	/**
 	 * 请求下载文件
 	 *
-	 * @param url      请求下载文件地址
-	 * @param timeout  超时时间
+	 * @param url     请求下载文件地址
+	 * @param timeout 超时时间
 	 * @return HttpResponse
 	 * @since 5.4.1
 	 */